win,fs: Upgrade `fs__access()` to respond to `X_OK`.
By utilizing the `AccessCheck()` ACL API, we can interrogate the
filesystem for executable permissions on a particular file for the given
process/user.  As this operation is slightly expensive, it is only done
if the user requests it by setting the `mode` parameter with `X_OK`.

In particular, this implementation checks for the `FILE_GENERIC_EXECUTE`
permission within the ACL entries of the given file.

(cherry picked from commit 1dcf32479840be96210d0807037351626cb82e2f)