win,fs: Add more complete `fs__chmod()` implementation
This implementation of `fs__chmod()` maps the familiar `owner`, `group`,
`other` triplet of permissions within a POSIX `mode` parameter to ACL
entries involving the current user, the current user's primary group as
well as any groups the user may belong to that already have ACL entries
within the given file object, and the `Everyone` group.  We create new
ACL entries explicitly allowing and denying the relevant permissions for
each of these security entities, and apply the new ACL to the given file
object.  The `chmod()` method continues to set the readonly bit upon the
given file when no write permissions are specified for any user.