win,fs: Use `AccessCheck()` for all `access()` calls
No longer shy away from using `AccessCheck()` for simple `W_OK`/`R_OK`
checks; use it for all calls unconditionally.

It turns out that `F_OK` is supposed to only check for path existence,
and because it is `0x00`, it ends up confusing the Windows APIs below
where it looks like we're asking for the null set of permissions, so
handle that as well.